clean and refactor tests
derrive test classes from TestSupport
reorganize tests (initial)
move compiler specific tests to own namespace
remove unused files / classes
etc etc
