+++ QueryBuilder.Tests/Oracle/OracleLimitTests.cs
@@ -1,16 +1,20 @@
-﻿using System;
-using SqlKata;
-using SqlKata.Compilers;
+﻿using SqlKata.Compilers;
+using SqlKata.Tests.Infrastructure;
 using Xunit;
 
-namespace SqlKata.Tests
+namespace SqlKata.Tests.Oracle
 {
-    public class OracleLimitTests
+    public class OracleLimitTests : TestSupport
     {
         private const string TableName = "Table";
         private const string SqlPlaceholder = "GENERATED_SQL";
 
-        private OracleCompiler compiler = new OracleCompiler();
+        private OracleCompiler compiler;
+
+        public OracleLimitTests()
+        {
+            compiler = Compilers.Get<OracleCompiler>(EngineCodes.Oracle);
+        }
 
         [Fact]
         public void NoLimitNorOffset()
