+++ QueryBuilder.Tests/SqlServer/SqlServerLegacyLimitTests.cs
@@ -1,15 +1,18 @@
-using SqlKata;
 using SqlKata.Compilers;
+using SqlKata.Tests.Infrastructure;
 using Xunit;
 
-namespace SqlKata.Tests
+namespace SqlKata.Tests.SqlServer
 {
-    public class SqlServerLimitTest
+    public class SqlServerLegacyLimitTests : TestSupport
     {
-        private SqlServerCompiler compiler = new SqlServerCompiler()
+        private readonly SqlServerCompiler compiler;
+
+        public SqlServerLegacyLimitTests()
         {
-            UseLegacyPagination = false
-        };
+            compiler = Compilers.Get<SqlServerCompiler>(EngineCodes.SqlServer);
+            compiler.UseLegacyPagination = true;
+        }
 
         [Fact]
         public void NoLimitNorOffset()
@@ -26,10 +29,7 @@ namespace SqlKata.Tests
             var query = new Query("Table").Limit(10);
             var ctx = new SqlResult {Query = query};
 
-            Assert.EndsWith("OFFSET ? ROWS FETCH NEXT ? ROWS ONLY", compiler.CompileLimit(ctx));
-            Assert.Equal(2, ctx.Bindings.Count);
-            Assert.Equal(0, ctx.Bindings[0]);
-            Assert.Equal(10, ctx.Bindings[1]);
+            Assert.Null(compiler.CompileLimit(ctx));
         }
 
         [Fact]
@@ -38,10 +38,7 @@ namespace SqlKata.Tests
             var query = new Query("Table").Offset(20);
             var ctx = new SqlResult {Query = query};
 
-            Assert.EndsWith("OFFSET ? ROWS", compiler.CompileLimit(ctx));
-
-            Assert.Single(ctx.Bindings);
-            Assert.Equal(20, ctx.Bindings[0]);
+            Assert.Null(compiler.CompileLimit(ctx));
         }
 
         [Fact]
@@ -50,11 +47,7 @@ namespace SqlKata.Tests
             var query = new Query("Table").Limit(5).Offset(20);
             var ctx = new SqlResult {Query = query};
 
-            Assert.EndsWith("OFFSET ? ROWS FETCH NEXT ? ROWS ONLY", compiler.CompileLimit(ctx));
-
-            Assert.Equal(2, ctx.Bindings.Count);
-            Assert.Equal(20, ctx.Bindings[0]);
-            Assert.Equal(5, ctx.Bindings[1]);
+            Assert.Null(compiler.CompileLimit(ctx));
         }
 
         [Fact]
